+++ spec/rubocop/cop/rails/delegate_spec.rb
@@ -120,6 +120,23 @@ describe Rubocop::Cop::Rails::Delegate do
     expect(cop.offenses).to be_empty
   end
 
+  it 'ignores delegation with assignment' do
+    inspect_source(cop,
+                   ['def new',
+                    '  @bar = Foo.new',
+                    'end'])
+    expect(cop.offenses).to be_empty
+  end
+
+  it 'ignores delegation to constant' do
+    inspect_source(cop,
+                   ['FOO = []',
+                    'def size',
+                    '  FOO.size',
+                    'end'])
+    expect(cop.offenses).to be_empty
+  end
+
   describe '#autocorrect' do
     context 'trivial delegation' do
       let(:source) do