+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## master (unreleased)
 
+### Bugs fixed
+
+* Fix `Delegate` for delegation with assignment or constant. ([@geniou][])
+
 ## 0.21.0 (24/04/2014)
 
 ### New features